// frontend/src/app/core/api/model/schedule-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ScheduleResponse { 
    id?: number;
    label?: string;
    classGroupName?: string;
    trainingName?: string;
    status?: ScheduleResponse.StatusEnum;
    classGroupId?: number;
    periodId?: number;
    periodLabel?: string;
}
export namespace ScheduleResponse {
    export const StatusEnum = {
        Draft: 'DRAFT',
        Active: 'ACTIVE',
        Archived: 'ARCHIVED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}