// frontend/src/app/core/api/model/active-class-group-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ActiveClassGroupResponse { 
    id?: number;
    name?: string;
    trainingType?: ActiveClassGroupResponse.TrainingTypeEnum;
}
export namespace ActiveClassGroupResponse {
    export const TrainingTypeEnum = {
        Continuous: 'CONTINUOUS',
        Accelerated: 'ACCELERATED',
        Accredited: 'ACCREDITED'
    } as const;
    export type TrainingTypeEnum = typeof TrainingTypeEnum[keyof typeof TrainingTypeEnum];
}