// frontend/src/app/core/api/model/default-period-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DefaultPeriodResponse { 
    id?: number;
    label?: string;
}